+++ lib/tubular/grid-page-info.ts
@@ -1,0 +1,8 @@
+export class GridPageInfo {
+    public currentInitial = 0;
+    public currentTop = 0;
+    public currentPage = 0;
+    public totalPages = 0;
+    public totalRecordCount = 0;
+    public filteredRecordCount = 0;
+}
